requirements.txt
absl-py==2.1.0
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.16.0
chess==1.10.0
cloudpickle==3.0.0
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
Farama-Notifications==0.0.4
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
grpcio==1.62.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
pycparser==2.22
pyparsing==3.1.2
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
Shimmy==1.3.0
six==1.16.0
stable_baselines3==2.3.0
stockfish==3.28.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tinycss2==1.2.1
torch==2.2.2
torch-tb-profiler==0.4.3
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
webencodings==0.5.1
Werkzeug==3.0.2
